Move functions to plan_bot.py
